Allow to include image size into avatar URL
For some avatar providers the preferred image size can be specified.
If the image size is not provided to the avatar provider it returns
the avatar image with the default size and the browser will do the
scaling. If the default image size is not square the scaling looks
ugly.

Change-Id: I77755bfe44bbd4856f6d459c6cf60b990112058e
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>